requirements.txt
databases[aiosqlite]
fastapi[all]
passlib[bcrypt]
python-jose[cryptography]
sentry-sdk[fastapi]
sqlalchemy
tinkoff-investments
